Refactored poll_cards.tsx
