+++ packages/commonwealth/client/scripts/views/pages/view_thread/ViewThreadPage.tsx
@@ -62,7 +62,8 @@ import { EditBody } from './edit_body';
 import { LinkedProposalsCard } from './linked_proposals_card';
 import { LinkedThreadsCard } from './linked_threads_card';
 import { LockMessage } from './lock_message';
-import { ThreadPollCard, ThreadPollEditorCard } from './poll_cards';
+import { ThreadPollCard } from './thread_poll_card';
+import { ThreadPollEditorCard } from './thread_poll_editor_card';
 import { SnapshotCreationCard } from './snapshot_creation_card';
 
 export type ThreadPrefetch = {
